pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pymodbus-vsensor"
version = "0.1.0"
description = "Communication with CMR Controls V-Sensor over Modbus"
authors = [{name = "", email = ""}]
readme = "README.md"
license = {text = "MIT"}
dependencies = [
    "pymodbus",
    "PyYAML",
    "streamlit",
    "paho-mqtt",
    "click"
]

[tool.pytest.ini_options]
addopts = "-q"
